requirements.txt
PyQt5
PyQt5-tools
pygame
mutagen
pillow





















































































































































































































































